+++ TerrainRender/TerrainRender/TerrainRender/Model/LineList.cpp
@@ -4,7 +4,7 @@
 #include	<memory>
 
 
-bool LineList::Initialize(Microsoft::WRL::ComPtr<ID3D11Device> device, IVertexShaderPtr vertexShader, IPixelShaderPtr pixelShader, VertexPolyLine* vertices, unsigned long* indices, UINT vertexCount, UINT indexCount)
+bool LineList::Initialize(Microsoft::WRL::ComPtr<ID3D11Device> device, IVertexShaderPtr vertexShader, IPixelShaderPtr pixelShader, VertexPolyLine* vertices, size_t* indices, size_t vertexCount, size_t indexCount)
 {
 	if (device == nullptr || vertexShader == nullptr || pixelShader == nullptr)
 		return false;
@@ -51,7 +51,7 @@ int LineList::GetVertexCount() const
 
 }
 
-bool LineList::InitializeBuffers(Microsoft::WRL::ComPtr<ID3D11Device> device, VertexPolyLine* vertices, UINT vertexCount)
+bool LineList::InitializeBuffers(Microsoft::WRL::ComPtr<ID3D11Device> device, VertexPolyLine* vertices, size_t vertexCount)
 {
 
 	D3D11_BUFFER_DESC						vertexBufferDesc;
